[tool: automated rendering] Restore Structurizr watermark + branding
I just feel like this is the right thing to do; Simon Brown added these elements to the diagrams rendered by Structurizr Express and it wouldn’t be right to remove them.

The only changes I’ve made here is to render the watermark in a fixed position every time, rather than a random position with a random rotation, because the randomness makes testing inconvenient (it means we can’t expect the output to be deterministic). Also added “and FC4” to the ”Powered by” line — seems reasonable.