[docs EP01] Overhaul Usage section and Workflows
This is a rather extensive set of changes, to both simplify and to
update the workflow in light of the new authoring workflow that was
recently shipped in #135.

We no longer want to do anything with the user’s clipboard, and we don’t
want to have two different workflows, one for editing a view and one
more editing a model. The workflow needs to be way simpler; you just
point fc4-tool at a repo and edit whatever you want, and fc4-tool does
the right thing.