[tool] Decomplect formatting/snapping: docs
Until now these features were complected as an “edit” or “process”
feature. This has been awkward and confusing for a good while — pretty
smelly. Eventually it dawned on me that I kept describing the feature
with an “and” or an “also” in there — and it became obvious that
(re)formatting the YAML and snapping the coordinates to a grid can and
should be discrete features.

This is also preparatory refactoring for #168.

This removes the CLI subcommand process as it was more trouble than it
was worth to refactor, and all the subcommands are going away soon as
per #168.

Refs #170