[tool] Decouple two more features: edit + watch
This is preparatory refactoring for #168.

As part of #168, we’ll be making the watch a mode that’s composable with
the main features, and we’ll be enabling users to toggle each of the
main features individually. In other words, we have three major
features: -f, -s, and -r. A user can invoke the tool with just one of
those — any one, or with any two of them, or with all three. They can
also optionally add -w to have the feature(s) applied in “watch mode”.

This is substantially different than the current paradigm, wherein a
specific and limited set of subcommands trigger a hard-coded “workflow”
or mode of operation.

In order to support the new approach, it is necessary to decouple the
watch feature from edit workflow — essentially, to extract the watch
feature from the edit workflow, and make it a higher-order function so
that it can be composed with any workflow (or featureset, mode of
operation, etc.) That’s what this change does.

Most of the code related to watching is moved from edit.clj to watch.clj
with minimal modifications. The biggest change is that the watch
workflow can no longer assume that if the function it invokes changes
the file that’d just been changed, that that change would be followed by
a rendering step that’d block for a substantial amount of time — at
least 300ms and often 1000–2500. And therefore we can no longer assume
that when processing subsequent events, that the file will still be in
the active set because it’s being rendered. Therefore I restored the map
in the Hawk “context” value that keeps track of when a file change event
was most recently triggered, for each given file. This had been removed
in a5274963eb because at the time it proved sufficient to simply keep
track of a current-active set and filter by membership in that set.

For more background on how the watch has worked to date, and avoided
duplicative work and infinite loops, see a5274963eb.

Anyway, since the tests now supply processing functions that are
extremely fast, and some of them change the subject file, I needed to
restore the time-based successive-event filtering. I kept the active-set
filtering for now, but I’m not sure it’s still needed. This is pretty
hard to reason about, honestly.