[tool] Add a :main profile to deps.edn
Makes it easier to run the tool from source, which is occasionally
helpful, maybe? I mean, I’m using it right now... but to test the CLI
UX. Arguably it’d be better to call -main from the REPL. I dunno,
sometimes you just want to run the tool from source. Even if it’s not
the most efficient thing.