[tool] Add CLI option --tmp-renderer
So as to enable users to opt-in to the experimental (pure Clojure)
renderer, as opposed to the default (Node) renderer.

The new renderer itself still has bugs, but that’s expected at this
point. In fact, that’s a big motivation for adding this option at this
stage: so we can get some help finding bugs.

Refs #159