[docs tool] Change recommended installation method to Homebrew
We just published a [Homebrew Tap][1] for Funding Circle’s FLOSS tools
and [that tap][2] includes a formula for fc4-tool.

This is great because not only does it make it _way_ easier to install
the fc4 tool, it also makes it _much_ more likely that users will
install updates!

(This _might_ also enable us to get ahold of some download/install
stats, but that might require contributing a formula to homebrew-core;
I’m not sure.)

[1]: https://docs.brew.sh/Taps
[2]: https://github.com/FundingCircle/homebrew-floss