[tool] Clear CircleCI dependency caches
This is necessary because the deps that are currently cached all lived
in directories that were relative to the user’s home directory — the
user that was used by default within the prior Docker image used to run
the tests — and that user happened to be root, which is of course
convenient but insecure.

Now, we’ve switched to a different Docker image, one wherein the default
user is *not* root — in this case, it’s circleci. But because the deps
haven’t changed, CircleCI is attempting to restore the deps to their
*prior* locations, rooted in /home/root/ — which was failing, because
the user circleci of course does not have access to /home/root

Therefore we need to clear the deps caches.

The technique used here was found here:
  https://circleci.com/docs/2.0/caching/#clearing-cache