[tool CI] Make some tests we run in CI more comprehensive
This CI job tests our distribution package, to ensure that it can be
used successfully to do real work. Since we recently added an
experimental renderer, and we’re about to make it the default (see
[this project][https://github.com/FundingCircle/fc4-framework/projects/2])
it seems like a good idea to test all these scenarios in this job.