[tool] Drop most flags used to launch browsers
I’m working in a branch to switch the Docker images we use in CI for
testing+building the tool from a custom image that we maintain to a
prebuilt image that CircleCI maintains ([docs][images]).

The tests are now passing just fine in that image, but the CI job that
tests the distribution packages is failing.

I re-started [one of those jobs][ci-job] with SSH enabled and managed to
reproduce the failure:

```
~/test/test-renders$ cat test-render-1.yaml | /tmp/fc4/fc4-render --verbose test-render-1.yaml > result-01.png
launching browser...
loading Structurizr Express from https://structurizr.com/express...
🚨🚨🚨
RENDERING FAILED: TimeoutError: Navigation Timeout Exceeded: 30000ms exceeded
    at _createTimeoutPromise.Promise.then (/snapshot/circleci/project/tool/renderer/node_modules/puppeteer-core/lib/NavigatorWatcher.js:94:21)
```

When I listed the files in that directory I saw a bunch of stack dumps
from Chrome, which seemed to imply that Chrome was crashing. And based
on the output of the renderer, it seemed to be crashing while loading
Structurizr Express (SE).

I then tried to reproduce this outside of the context of fc4 tool, so as
to perhaps enable some logging, like so:

```
$ google-chrome --headless --disable-gpu --print-to-pdf https://structurizr.com/express/
```

but that seemed to succeed — it didn’t crash, at least, and it produced
a PDF file.

After a while, I realized that I should try passing all the same flags
to Chrome that the Node renderer is using when launching the browser (as
found in render.js) so I added `--no-sandbox` and
`--disable-dev-shm-usage`:

```
$ google-chrome --headless --no-sandbox --disable-dev-shm-usage --disable-gpu --print-to-pdf https://structurizr.com/express/
Fontconfig warning: "/etc/fonts/fonts.conf", line 100: unknown element "blank"
[0809/171407.252069:FATAL:gpu_data_manager_impl_private.cc(892)] The display compositor is frequently crashing. Goodbye.
Failed to generate minidump.Illegal instruction (core dumped)
```

Aha! So the crash is related to using one, or both, of those flags.

I then tested the command with each flag individually, and found that
the crash occurred when specifying only `--disable-dev-shm-usage` but
did not occur when specifying only `--no-sandbox`.

So it would appear that `--disable-dev-shm-usage` is the culprit.

This is surprising, because that flag is [recommended][pptr-docs] by the
Puppeteer team when running Chrome in Docker — but whatever, it works!
¯\_(ツ)_/¯

While I was removing that flag, I realized we can also remove
`--no-sandbox` because in the new Docker image in which we’re running
all the tests, the default user being used is no longer root (it’s now
`circleci`).

[ci-job]: https://circleci.com/gh/FundingCircle/fc4-framework/2167
[images]: https://circleci.com/docs/2.0/circleci-images/
[pptr-docs]: https://github.com/GoogleChrome/puppeteer/blob/master/docs/troubleshooting.md#tips