[tool] Upgrade puppeteer
Another change related to switching to a new test/build Docker image.

The new base image has a newer version of Chrome than the prior image
(which was Chromium rather than Chrome but that *probably* doesn’t
matter) and it’s always best to use a version of Puppeteer designed for
the specific version of Chromium one is using. (When one is using
Chrome, one is using Chromium (but not vice-versa).)