[tool] Small code cleanup related to setting a system property
I happened to be scrolling by this code and had a sudden epiphany that
I could just set this system property when launching the JVM instead of
in the code before requiring/importing certain libraries/classes.

So, if this works, this should be nice.

Just for reference, I found this approach to preventing the Java app
icon from popping up and grabbing focus on MacOS here:
  https://stackoverflow.com/questions/17460777/stop-java-coffee-cup-icon-from-appearing-in-the-dock-on-mac-osx/17544259#comment48475681_17544259