[tool] Fix the CI job that publishes the packages…
…the distribution packages for new releases, that is.

This was broken by #199, and we didn’t know it was broken until after we
merged because this job is only run on master.

The reason it failed is because this job doesn’t include a checkout
step, so `git rev-parse` was failing with fatal: Not a git repository
because the dir `.git` wasn’t present.

Example of a failed job:
  https://app.circleci.com/jobs/github/FundingCircle/fc4-framework/2564

I could have fixed this by adding the checkout step to the job, but that
would have made it slower. I chose instead to use the CIRCLE_SHA1
environment variable, shortened to 7 chars; this results in the same
value as `git rev-parse --short HEAD`.

I also changed the job that creates the distribution packages to use the
same expression, for consistency. It’s unlikely, but the behavior of
`git rev-parse --short` *could* change.