[tool] Minor reorg of renderer code
This is preparatory refactoring for some subsequent commits that will
add code related to conjoining SVG diagrams and keys together. I decided
that chromium-renderer.clj was just getting too big.

I figured at the same time I’d improve the naming of the namespaces and
the record a little to make them shorter and/or clearer.

Refs #222