Minor tweaks to CI configs
I recently read a post on the CircleCI blog:

https://circleci.com/blog/optimizing-react-native-camera/

and one of the tips I picked up here was to save the dependency cache
right after downloading/installing dependencies, rather than at the end
of a job. This made sense to me:

> This is not optimal because saving the cache will not happen if any of
> the other steps in the job fail before the save_cache step has a
> chance to execute. It’s best to place the save_cache step immediately
> after a dependency step - this guarantees the cache will be saved as
> long as the dependency step works...

I also noticed that one of the jobs (tool_eastwood) was using a Docker
image with Clojure preinstalled, but then installing a more recent
version of Clojure, because the version in the CircleCI Docker image was
too old for Eastwood. So I changed the job to run in CircleCI’s openjdk
image instead, just to make the job hypothetically a little lighter and
a little simpler.

Now that I’m thinking about this, the version of Clojure that we’re
using via the circleci/clojure repository is unstable, because we’re not
pinning the jobs to a specific version of the image. Since almost all
the jobs are now using circleci/openjdk and installing a specific
(stable) version of Clojure via bin/install-clojure, I think it’d make
sense to just to that across the board, for all the jobs. So I’m doing
that now.