Enable rendering on Arch Linux
In #236, @mathpunk reported that when they attempted to use FC4 on Arch
Linux, they got a compilation error.

The prior commit helped me reproduce and debug this; it turns out the
main problem is that the function `chromium-path` is really dumb right
now — it’s mostly a hard-coded list of paths. That list was missing the
install path for Chromium on Arch. So I added it.

Once I did that, FC4 still couldn’t launch Chromium; it would exit with
status code 0 and print this error message:

> Running as root without --no-sandbox is not supported.

That was familiar. The function `chromium-opts` builds a vector of CLI
options (flags) to include when invoking the browser, and that function
*used* to include --no-sandbox in its return value, but I removed it at
one point — I _think_ it was when we adopted the CircleCI Docker images,
which, unlike most Docker images, do _not_ set the user to root.

So I restored that option, and once I did that, FC4 was able to launch
Chromium and render a diagram (in the local Docker container in which I
was manually testing).

Assuming the tests pass in CI, that should be #236 fixed.

----

I want to keep this commit as small and simple as possible, but some
follow-up improvements suggest themselves:

* `chromium-path` should be a lot smarter
  * e.g. instead of, or in addition to, a list of static paths to check,
    it should probably also attempt to locate the browser via
    introspection, e.g. `which` or somesuch
* When `chromium-path` returns `nil`, `start-browser` should return an
  anomaly, which should result in a useful, informative, and actionable
  error message being printed to the terminal (written to stdout).
* After `start-browser` attempts to launch the browser, it should check
  that the browser actually launched successfully.
  * And if it didn’t, then the function should return an
    anomaly, which should result in a useful, informative, and
    actionable error message being printed to the terminal (written to
    stdout).

I actually had started on that last one back in July, in a local
branch named `check-browser-launch`. I’ll push this to the repo now,
just so it’s visible to more folks than just me.

Fixes #236