[tool] Render SVG diagrams to .svg files
As explained in #222:

> In #200 we added support for rendering diagrams to SVG images instead
> of, or in addition to (simultaneously with) PNG images.
>
> When I implemented that, I had the tool save the SVG images as files
> with an `.html` extension, because that was what Structurizr Express
> (SE) did. I didn’t know why that was the case, but I trusted that
> there was a good reason for that choice, so I figured I’d do the same
> thing. After all, when I opened those HTML files — both those exported
> directly from SE and those created by fc4 — in a browser — they
> rendered nicely. So it seemed good to me.
>
> Unfortunately, I neglected to account for an important goal of FC4: to
> enable peer review of changes to diagrams via GitHub Pull Requests,
> via textual diffs of the YAML sources and via visual diffs of the
> rendered images. GitHub will render SVG files with the `.svg`
> extension (as long as their contents are valid) but it won’t render
> any files with the `.html` extension.
>
> This is bad. So let’s fix the tool to save the SVG image files as
> valid SVG with the `.svg` filename extension.

This was a little more involved than one might think. I think it’s
because browsers are way more tolerant of strange issues in HTML files
as compared to SVG files.

One result of that was that the HTML file worked without a root element
— there were two SVG documents at the “root” of the file. That’s not
valid for XML, so in order to create a single SVG XML file, I had to
create a new root element and wrap the diagram and the key in SVG `g`
elements (`g` for _group_). Anyway, it works.

Fixes #222