Fix CI script install-sys-deps
In e2d53c904 I attempted to enhance this script, but I introduced a bug
into an error case (when no args are supplied) and forgot to test that
case.

This is a good example of why all scripts should be thoroughly tested,
even “support” scripts.

In this case, this isn’t a user-facing script, it’s only meant for use
within our CI environment, so I guess I can live with it not being
thoroughly tested. That said, we might want to consider moving scripts
specific to a given CI environment/config closer to the config.
i.e. perhaps we should move this script into `.circleci/bin`