Delete model, view, and styles DSLs
These data models and the associated code are not currently in use, and
as per #108 I have a significant set of changes coming soon that will overhaul/refactor/enhance these data models.

I took a look at those changes, with their massive diff, and realized
that it’d be quite challenging for anyone currently involved with FC4 to
review the diff, because it’s massive, and because no one currently
working on FC4 has any context on this code — they’d have to first learn
it before they could make sense of the diff.

So as a form of preparatory refactoring, I’m just deleting this version
of this code now, and will re-introduce it in a new branch+PR within a
day or two.

Refs #108