Move linting configs out of the project root
Now that we’ve moved a bunch of stuff from `tool` into the root of the
repo (in #254 / f6e35ae) there’s too much stuff in the root.

There are two problems with having so much stuff in root:

1. It makes it hard to scan the root and get a sense of what’s there and
   why.
   1. This is true for me, so I suspect it’d be doubly true for less
      experienced devs or devs who will be new to this project.
1. In the repo’s “home page” in GitHub, it pushes down the content of
   the readme farther “below the fold” so people have to scroll farther
   to get to the content — and they might not bother.
   1. This is not a huge deal, but it does still matter to me.

This commit just moves some linting-related config files out of the
root, which isn’t really significant, but there’s more in the next
commit. I just wanted to do this in a few smaller chunks.