Fix flaky tests on MacOS in CircleCI
This test was intermittently failing with output like this:

```
a body that takes longer than the timeout to finish
expected: (thrown?
           TimeoutException
           (u/with-timeout 100 (do (Thread/sleep 150) true)))
  actual: nil
```

And sometimes the inverse.

I assume this has something to do with the virtualization setup, or the
fact that the MacOS executors on CircleCI are seemingly slow in general.