[website] Add Jekyll config file
I added this file because on my laptop the github-pages gem was blocking
jekyll from building the site. Turns out _that_ happened because I’d
renamed the “remote” in my local git repo from `origin` to
`gh-fundingcircle` because I decided “origin” is counter to the ethos of
git. But then it turned out that github-pages had been using that
default remote name to determining the GitHub repo for this local repo,
and then retrieving information about the repo from GitHub (I think), or
at least using the remote URL to extract the repo name and thereby the
default site title.

ANYWAY, so that’s why I created this config file, and that’s why it
includes the `repository` property.

I added the `title` property because I didn’t like the site title that
the github-pages gem was setting automatically, which was a version of
the repo name at GitHub (fc4-framework). I didn’t like that mainly
because I’m in the process of reframing FC4 as a tool rather than a
framework (#226).

Refs #226