Update tests to handle changes to Structurizr Express
I was working on something unrelated and the rendering tests started
failing with output like this:

expected: (<= distance-percentage (:svg max-allowable-image-differences))
  actual: (not (<= 0.052500776638707676 0.05))

as you can see, the rendered SVG image was 5.25% different than the
expected image, which exceeded the max allowable difference of 5%.

I took a look and it seems that Simon Brown has made some minor changes
to Structurizr Express’ rendering; mostly to the watermark that
Structurizr Express has been adding to rendered images. Well, actually
he seems to have removed the watermark and replaced it with a larger,
more opaque Structurizr wordmark in a fixed location.

This is of course 100% fine; it’s his product and it’s his prerogative
to change it however he sees fit, at any time. I have zero problems with
him making this change unannounced; when I chose to have FC4 “wrap” SE
and invoke its rendering routines programmatically, I knew that there
were zero guarantees that SE would continue to function, or to be
available, and I knew that it would be my responsibility to update FC4
if and when SE changed or was retired.

So! There are two changes here:

1. Update the “expected” image to one rendered with the new wordmark
2. Increase the maximum allowable image difference to 6%

I made the latter change because the differences in the rendering here
were so visually trivial, to my eye, that it wasn’t necessary for the
tests to fail in this case.