Website: fix #228: squished image
As per #228:

> When the website is viewed on a phone the example diagram is stretched
>
> ...
>
> Not just on a phone, actually; this is easily reproducible in a WIMP
> computer or some tablets by making the window very narrow.

As for the fix... TIL about modern html/browsers: the `width` and/or
`height` attributes of `img` elements no longer instruct browsers of the
dimensions to use when rendering the image. Rather, they inform browsers
of the “intrinsic size” of the image, which sets the space taken up by
an image to ensure the page layout is stable before it loads. ([MDN][1])

To instruct browsers of the dimensions to use when rendering the image,
we apparently need to use the CSS properties `min-width`, `max-width`,
`min-height`, and `max-height`. I would have thought that specifying
just `width` and/or `height` would have worked, but nope.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Img

Fixes #228