Send coverage data to Coveralls
I’ve been frustrated with Codecov for awhile so I want to try this
alternative. It was easy to figure out how to generate the JSON file for
Coveralls but not at all easy to figure out how to POST the file to
the Coveralls API. Their documentation needs some work, which does not
inspire much confidence.