Website: fix links and redirects
When I tested the website locally, the absolute links (those starting
with a slash) worked just fine, because in that scenario the site is
“rooted” at the host root — e.g. `docs/manual/scheme.md` is
accessible at `http://localhost:4000/docs/manual/scheme.md` so an
absolute link like `/docs/manual/scheme` works just fine.

When the website is published to GitHub Pages, however, it’s a
[“repo site”][gh-pages-types] — so it’s rooted under a subdirectory,
i.e. there’s an additional segment in the URL, the first one, so
`docs/manual/scheme.md` becomes
`https://fundingcircle.github.io/fc4-framework/docs/manual/scheme.md` so
an absolute link like `/docs/manual/scheme` won’t work — it has to be
relative so that it’ll work in either scenario.

So the links all broke when we merged #257 a few minutes ago.

So I changed all the absolute links to relative.

This also made me realize that the way I’d constructed the redirects
might not work, and even if they did work, they didn’t necessarily
sense. So I changed those to use relative paths as well, which is more
consistent and sensical, and I also removed the redundant redirects, as
[jekyll_redirect_from][jrf] just created HTML files that direct the
browser to redirect to the new URL (using HTML tags, not HTTP status
codes). In other words, there never was a URL like `tool.html` — there
was only `tool/index.html` which was also accessible via `tool` or
`tool/`

[gh-pages-types]: https://git.io/JepJd
[jrf]: https://github.com/jekyll/jekyll-redirect-from