Bifurcate CircleCI workflow
I’m working on adding yet another job to the workflow, and I saw that
the workflow had become massive, and realized that it didn’t need to be
so big; that there was a natural seam on which to split the workflow in
two so as to make the pipeline* more tractable.

This is preparatory refactoring to add another CI job in the next
commit.

* “pipeline” is a new (ish) CircleCI term for a set of workflows. As far
  as I can tell, a CircleCI config.yaml file defines and corresponds to
  a single pipeline.