Address PR feedback: better name for a spec
I had named this spec using a sort of class-oriented conceptual model,
which made it hard to talk about how/where/why it was used. So during PR
review we decided to adopt this conceptual model instead. It’s inspired
by how HTML elements have a shared set of common elements in the HTML
spec.

Addresses feedback on #251.