Fix: CI job publish_dist_pkg fails on master
This is a follow-up to #266 / 99f481bc1b. In that commit I modified the
CI job publish_dist_pkg to always run on master. Unfortunately I
neglected to do the same for its dependencies: build_dist_pkg and
test_dist_pkg. So publish_dist_pkg did indeed run on master after that
PR was merged, but it failed, because the distribution package hadn’t
been created. Doh!