Always run all CI jobs on master
This is a follow-up to #266 / 99f481bc1b. In that commit I modified the
CI job publish_dist_pkg to always run on master. Unfortunately I
neglected to do the same for its dependencies: build_dist_pkg and
test_dist_pkg. So publish_dist_pkg did indeed run on master after that
PR was merged, but it failed, because the distribution package hadn’t
been created. Doh!

At first (22b7cd1) I modified the CircleCI config so that those specific
jobs would always run on master. But then I decided it would be simpler,
probably be safer, simpler, clearer, and easier to debug and
troubleshoot, and more concise, to simply always run all CI jobs on
master. This is also more consistent with how the CI config has worked
for the past year+. So I reverted that commit (fead3a54) and went this
this change instead.