Attempt to speed up our CircleCI workflow “test”
Since we merged #251 yesterday, executions of our CircleCI workflow
“test” on master and subsequent branches have been very slow; the last
execution on master was ~11 mins.

This branch just makes the situation worse; the last execution of the
workflow “test” on this branch took ~16 minutes. This is bad.

At first I tried scaling up our CI “machines” via the CircleCI executor
config property `resource_class` but that didn’t work; it turns out that
on our current CircleCI plan we have access only to the default `medium`
class and to `small` — but nothing like `large` or above.

Then it occurred to me that it might be time to change the
multithreading strategy we’re using with our test runner, eftest. It’s
been set to `:vars` for a long time, because way back then when I first
set it, that was the setting that yielded the fastest results.

That setting is described like so in the eftest docs:

> If you want the test vars inside a namespace to be executed in
> parallel, but the namespaces themselves to be executed in serial, then
> set the `:multithread?` option to `:vars`

When I thought about this, especially in light of the changes introduced
in #251 and in this branch, *and* in light of our constrained CircleCI
compute resources (only 2 vCPUs), I thought — yeah, this is probably a
bottleneck. See, in #251 and in this branch, we introduced three new
tests, in a single test namespace — `fc4.dsl.model-test` — all three of
which are very CPU-intensive. And with the `:vars` setting, all three of
those tests would be running concurrently, and probably causing lots of
contention between each other’s threads for CPU time.

That’s why I thought it might be worth a try to switch to `:namespaces`
which is explained like so in the docs:

> If you want the vars inside a namespace to execute in serial, but the
> namespaces to be executed in parallel, set the `:multithread?` option
> to `:namespaces`

I think this means that the three tests in this namespace will *never*
run concurrently. Which might yield some improvements. Let’s see!

eftest docs: https://github.com/weavejester/eftest/#multithreading