Set thread count of test runner
See the diff for the reason.

I tested this by running the tests in `fc4.dsl.model-test` five times on
the parent commit and then five times with these changes. I made sure
that the Docker container I was using had 2 vCPUs so it would be similar
to our Linux test containers on CircleCI.

The results:

| commit             | times (wall-clock seconds) | median |
| ------------------ | -------------------------- | ------ |
| Parent (7b7205a72) | 23.3 32.4 22.2 27.4 23.5   | 23.5   |
| This commit        | 21.7 23.5 24.6 19.7 21.9   | 21.9   |

Hmm, that’s not a huge improvement, but 7% isn’t bubkes — I’ll take it.