Fix bug in CI script halt-if-unchanged
I noticed that this script was halting in cases when it shouldn’t have
been — on this branch, actually.

It was the new job added in 72fb2fa298 — it wasn’t running past the call
to halt, even though it should have.

The step in that new job:

```
- run: .circleci/bin/halt-if-unchanged .circleci src deps.edn
```

As you can see, it’s meant to halt if all those paths are unchanged in
the current branch. The thing is, in *this* branch, it should *not* have
been halting, because in this branch, in fact, in *that* commit, there
were in fact lots of changes under `src`.

It occurred to me that perhaps this script is really only looking at
changes in the most recent commit in the branch, rather than all the
commits in the branch. That certainly seemed the case. So I reread the
script and tested it manually, and yeah, I confirmed it.

So I did a little reading and found an approach that’s simpler and
seems to actually work: diffing the current branch against master, but
only for the specified paths, and checking whether the diff is empty.

So let’s give this a try!