Add minor missing step to a CI job
I noticed that the other test jobs included a step to halt if the branch
doesn’t include changes to the code, but this one didn’t, so I added it,
mostly for consistently but also I guess to save a little time (even
though linting isn’t on the critical path for our CI pipelines
wall-clock execution time).

Also for consistency I swapped the order of two steps in one of the test
jobs. Now all three test jobs have the same initial steps in the same
order.