Add (failing) tests of the new model loading feature
…which was merged earlier this week via #23 / cd25ebe04e.

I was reviewing some of this code while resolving some merge conflicts
resulting from rebasing a local branch onto master, and I noticed that
a few things were off in some of the specs in `src/fc4/dsl/model.clj`
and `src/fc4/io/dsl.clj`.

I’ll explain the details in the next commit wherein I’ll fix the
problems. *This* commit adds some missing tests for the functions in the
latter file — tests that are failing, as of this commit, but will be
fixed by the next — and also activates instrumentation for all the tests
of the namespace, just for greater rigor.