Fix some failing tests related to model loading
In the prior commit I enabled clojure.spec instrumentation within the
test ns `fc4.io.dsl-test` to make the testing more rigorous. This
immediately caused the previously-existing test to start failing,
because the paths that were being supplied to the function under test as
args didn’t conform to the spec for the arg, as specified by the fdef
for the function. This fixes those tests by fixing those args such that
they now conform to the specs.