Copy the website from the main branch
I recently changed the default branch of the GitHub repo from `master` to `main` and it
seems like this may have inadvertently unpublished the project’s GitHub Pages website.

So this is an attempt to re-publish the site. I’ve copied the files from the dir `docs` of
the main branch, and I’m committing them here to this new “orphan” branch `gh-pages`. It
seems worth a try, at least.

If you have any questions, concerns, or objections about renaming the default branch,
please see:
https://seankilleen.com/2020/06/how-to-rename-your-default-branch-away-from-master-while-using-github-pages/