Refactor the website
Until now we’ve had GitHub Pages configured to publish our website from
the `docs` dir in the default branch of the repo. That worked fairly
well (except for having to explain why `docs` contained a dir named
`docs`) but I broke it when I renamed our default branch from `master`
to `main`. It appears that GitHub Pages is currently hard-coded to use
a branch named `master` as the publication source, rather than the
default branch, regardless of its name.

So, I created a new “orphan branch” named `gh-pages` and copied
contents of `docs` to the root of the project, in that branch. That’s
now working fine as the publication source for GitHub Pages, so the site
is back up.

Now that we’re no longer publishing from the default branch, we can
rename this dir. Its name was always annoying, especially the nested
`docs` dirs which was really confusing to explain.